Added pin assignments for UART5, LSF0204 enable
Removed (commented out) mentions of TNC-Black UARTS; added mentions (not in use) of pins allocated for eMMC and HDMI. It turns out that P8_17 and P8_18 (I2C1 SCL and SDA, respectively) are used by the Camera cape, as are P9_19 and P9_20 (I2C2 SCL and SDA, respectively). Need to determine if this means I2C is off limits or can co-exist on either interface.