Create fixtelemetry.plx
Until the logs spit out properly formatted JSON, this file serves as a template for getting telemetry information out of the "big" log as it currently gets written. It writes a file, "telemetryoutput.json", that can be read by jsonlog2.py, found elsewhere in this folder. It'll probably break pretty frequently, but I'll try to keep up with the changes.