Added pin & updated names; correction to P9_31
Improved names for Modem pins (there are multiple conventions for naming pins); added the power-on-or-key-up pin for the TX2H radio. Note that, in our experience, that radio doesn't behave well if you key it every x seconds; you pretty much have to power it up and leave it transmitting, or else you can get the unit locked in a half-state or something, with a locked-key-down transmit but low power, or something.